close to search working
