+++ package.json
@@ -17,7 +17,7 @@
     "react-dom": "^18.2.0",
     "react-icons": "^4.6.0",
     "react-router-dom": "^6.4.2",
-    "react-scripts": "5.0.1",
+    "react-scripts": "^5.0.1",
     "scripts": "^0.1.0",
     "styled-components": "^5.3.6",
     "web-vitals": "^2.1.4"
