+++ src/Components/dataContext.js
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const DataContext = createContext("")
